chore: return last fetched price in certain conditions + removal of some logs
Sometimes, the RPC does not have access to the historical state of the
price in the block in which is fetching. For these cases, an error
emerges. There's also another case where no routing through the
connectors has been discovered, therefore the contract returns 0 price.
For both of these situtations, the code now returns the last known price
of the token (i.e it simply doesn't update it, just the timestamp is
updated)